move http_fs_get
