+++ applications/system/curl/curl.c
@@ -6,14 +6,17 @@
 #include <cli/cli_status.h>
 
 #include <mongoose.h>
+#include <mongoose_glue.h>
 #include <wifi/wifi.h>
 #include <network/network.h>
 
 #define TAG "Curl"
 
 #define CURL_CA_BUNDLE_PATH "/ext/ca_bundle.crt"
+#define CURL_USER_AGENT \
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/138.0.0.0 Safari/537.36"
 
-// #define CURL_DEBUG
+//#define CURL_DEBUG
 
 #ifdef CURL_DEBUG
 #define CURL_INFO(...)  FURI_LOG_I(__VA_ARGS__)
@@ -30,15 +33,9 @@ typedef struct {
     FuriString* url;
     int event;
     FuriString* response;
+    uint8_t* data_body;
+    size_t data_body_size;
 } Curl;
-
-const struct mg_fs* http_fs_get(void);
-
-// int mg_url_is_ssl(const char *url) {
-//   return strncmp(url, "wss:", 4) == 0 || strncmp(url, "https:", 6) == 0 ||
-//          strncmp(url, "mqtts:", 6) == 0 || strncmp(url, "ssl:", 4) == 0 ||
-//          strncmp(url, "tls:", 4) == 0 || strncmp(url, "tcps:", 5) == 0;
-// }
 
 static void curl_mg_handler(struct mg_connection* connection, int event, void* event_data) {
     UNUSED(event_data);
@@ -50,24 +47,37 @@ static void curl_mg_handler(struct mg_connection* connection, int event, void* e
 
         if(mg_url_is_ssl(furi_string_get_cstr(instance->url))) {
             const struct mg_tls_opts opts = {
-                .ca = mg_file_read((struct mg_fs*)http_fs_get(), CURL_CA_BUNDLE_PATH),
-                .name = name};
+                .ca = mg_file_read(http_fs_get(), CURL_CA_BUNDLE_PATH), .name = name};
             mg_tls_init(connection, &opts);
         }
 
         mg_printf(
             connection,
-            "GET %s HTTP/1.0\r\nHost: %.*s\r\n\r\n",
+            "GET %s HTTP/1.0\r\nHost: %.*s\r\nUser-Agent: %s\r\n\r\n",
             mg_url_uri(furi_string_get_cstr(instance->url)),
             name.len,
-            name.buf);
+            name.buf,
+            CURL_USER_AGENT);
 
     } else if(event == MG_EV_HTTP_MSG) {
         struct mg_http_message* hm = event_data;
         CURL_INFO(TAG, "Data received: %.*s", (int)hm->message.len, hm->message.buf);
 
-        furi_string_printf(instance->response, "%.*s", (int)hm->message.len, hm->message.buf);
+        furi_string_printf(instance->response, "%.*s", (int)hm->head.len, hm->head.buf);
         instance->event = event;
+
+        if(hm->body.len) {
+            instance->data_body = malloc(hm->body.len + 1);
+            instance->data_body_size = hm->body.len;
+            memcpy(instance->data_body, hm->body.buf, hm->body.len);
+#ifdef CURL_DEBUG
+            FURI_LOG_I(TAG, "Header (%lu bytes):", (uint32_t)(hm->body.len));
+            for(uint32_t i = 0; i < hm->body.len; i++) {
+                FURI_LOG_RAW_I("%02X ", instance->data_body[i]);
+            }
+            FURI_LOG_RAW_I("\r\n");
+#endif
+        }
 
         connection->is_draining = 1;
         instance->done = true;
@@ -93,9 +103,9 @@ static Curl* curl_alloc(FuriString* url) {
 
     instance->network = furi_record_open(RECORD_NETWORK);
     network_init_current_thread(instance->network);
-
+#ifdef CURL_DEBUG
     mg_log_set(MG_LL_VERBOSE);
-
+#endif
     mg_mgr_init(&instance->mgr);
     mg_http_connect(
         &instance->mgr, furi_string_get_cstr(instance->url), curl_mg_handler, instance);
@@ -111,6 +121,10 @@ static void curl_free(Curl* instance) {
 
     furi_string_free(instance->url);
     furi_string_free(instance->response);
+    if(instance->data_body) {
+        free(instance->data_body);
+        instance->data_body = NULL;
+    }
     free(instance);
 }
 
@@ -128,7 +142,7 @@ void curl_url(PipeSide* pipe, FuriString* url, FuriString* args, void* context) 
 
     Curl* instance = curl_alloc(url);
     const char spin[] = "|/-\\";
-    int i = 0;
+    uint8_t i = 0;
     while(!instance->done) {
         mg_mgr_poll(&instance->mgr, 1000);
         printf("\rWaiting... %c", spin[i]);
@@ -140,6 +154,9 @@ void curl_url(PipeSide* pipe, FuriString* url, FuriString* args, void* context) 
     if(instance->event == MG_EV_HTTP_MSG) {
         printf(ANSI_FG_GREEN "Request succeeded\r\n" ANSI_RESET);
         printf("%s\r\n", furi_string_get_cstr(instance->response));
+        if(instance->data_body && instance->data_body_size) {
+            printf("%.*s\r\n", (int)instance->data_body_size, instance->data_body);
+        }
     } else if(instance->event == MG_EV_ERROR) {
         printf(ANSI_FG_RED "Request failed\r\n" ANSI_RESET);
         printf("%s\r\n", furi_string_get_cstr(instance->response));
