+++ targets/f20/mongoose/mongoose_glue.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <mongoose.h>
+
+struct mg_fs* http_fs_get(void);
